Some reorg, pulled the single hex move function into its own file.
